* unsigned int → int_fast16_t

* uint→int; < → !=

* unsigned int → int16

* Check nTip immediately

* const int16s

* Check that tree's not too large

* Stay within array limits (!)

* t1_descendantsof on heap, not stack

15% faster and no overflow concerns!

* n_internal

* M → heap

Slightly slower, but allows larger vals?

* least16 faster than fast16 (!)

* MAX_NODE & document

+++ src/tree_distance_functions.cpp
@@ -57,8 +57,8 @@ double lg2_trees_matching_split (int a, int b) {
  * nkK is converted to pkK in the calling function, when the sum of all
  * elements is divided by n.
 */
-double ic_element (const double nkK, const unsigned int nk,
-                   const unsigned int nK, const double n) {
+double ic_element (const double nkK, const int16 nk,
+                   const int16 nK, const double n) {
   if (nkK && nk && nK) {
     return nkK * log2(nkK * n / double(nk * nK));
   } else return 0;
