+++ src/tree_distances.h
@@ -1,6 +1,6 @@
 #include <limits>
-#include <stdint.h>
 #include <Rcpp.h>
+#include "ints.h"
 
 using namespace Rcpp;
 
@@ -9,8 +9,8 @@ using namespace Rcpp;
 
 typedef uint_fast64_t splitbit;
 typedef int_fast64_t cost;
-typedef int lap_row;
-typedef int lap_col;
+typedef int16 lap_row;
+typedef int16 lap_col;
 
 /*************** CONSTANTS  *******************/
 
@@ -41,13 +41,13 @@ extern double lg2_double_factorial[MAX_TIPS + MAX_TIPS - 2],
 
 extern int count_bits (splitbit x);
 
-extern cost lap(int dim, cost **assigncost,
+extern cost lap(int16 dim, cost **assigncost,
                 lap_col *rowsol, lap_row *colsol,
                 cost *u, cost *v);
 
 extern double lg2_trees_matching_split(int a, int b),
-  ic_element (const double nkK, const unsigned int nk,
-              const unsigned int nK, const double n),
+  ic_element (const double nkK, const int16 nk,
+              const int16 nK, const double n),
   one_overlap (const int a, const int b, const int n),
   one_overlap_notb (const int a, const int n_minus_b, const int n),
   spi (const splitbit* a_state, const splitbit* b_state, const int n_tips,
